Make postproc testable
This change moves much of the sourcing of data and parameters outside the calculate_average_emissions module, and into the user_driver.post_process method. This becomes an integration point with the greater system while calculate_average_emissions becomes more pure.

This then allows for an integration test that provides paths to test data where we can enshrine the current behaviour in tests, to protect against unintentional changes in the future.