Add manual_postproc.py script
This allows post processing to be done at will on a posterior multipliers solution file, or on the last successful iteration of a monthly fourdvar run.

To allow post processing to run on the physical data or on data from a file, loading and normalising the data has been made more generic.