Set stdout and stderr output logging when running via mpirun
The mpirun command will execute the command on multiple cores simultaneously. This makes it tricky to get access to stdout and stderr, since there will be a stdout and stderr stream for each concurrent process.

The -outfile-pattern and -errfile-pattern args will instruct mpirun to write these streams to files with the node number so that all output can be examined.