Add 'daily' sync type to load_from_archive.py
While the default behaviour remains the same, loading all the daily results for a monthly run, this adds a new 'daily' mode which syncs a single day into the working root. This will allow syncing already processed data from a previous run of the same day, if one exists. If no data exists because the daily run has never been previously run for the current date, the script will skip that day.