Remove archive loading from alerts_baseline.py
This has the unfortunate side-effect of tying the operation of this command tightly to our infrastructure, and makes it more difficult to run the logic separate from loading the data.

In it's place, we'll plan to add another load_from_archive step to the production workflow to ensure the data is present when we run the alerts_baseline script in production.