Add the archive-success task to docker-e2e-monthly.sh
For the most part, we probably want to leave the archive-load and archive-success tasks commented out in this script, since we don't want to add a dependency to S3 to this local workflow. However, they have been added and tested in case we do want to test them locally when making changes.