Raise exceptions in tropomi_methane_preprocess.py so jobs fail
Going back through git history I can't find an explanation for why all exceptions during obs preprocessing a file are ignored. There is currently an issue in this task in the prod workflow causing preprocessing to fail, but the workflow marks the job successful because it exits with a 0 status code.

This change will cause the script to exit with a non-0 status, so the workflow marks the job as failed. This will allow us to redrive the workflow execution and possibly recover from the failure.

In the worst case, it will properly alert us when this task has failed to process observations so that we can diagnose and fix any root causes.