Fix empty STORE_PATH when archive.py is run via eventbridge
The previous changes to source STORE_PATH from env variable assumed that archive.py would only be called from a workflow step, and failed to understand why the path was being re-constructed in the first place.  This script is also called when a workflow fails, but in this context will not have STORE_PATH set.

Failed jobs have been archiving the entire EFS aust10km folder, leading to a large amount of duplication and partial archives due to timeouts.