Update load_from_archive helpers to support new alert baseline workflow
The monthly workflow will be changed to always produce an alerts baseline file that covers the entire month, which will be useful for diagnostics. To facilitate this, the simulated and satellite observations from each daily run will need to be loaded as part of the monthly workflow.

Preparing an alerts baseline across a larger period that can be used in the daily alerts process will be spun out as a separate workflow, which will need to gather a smaller amount of data across a configurable period. Instead of specifying a number of days and an end date, this will now use the same START_DATE and END_DATE as other workflows.