Refactor naming of s3 buckets in archive scripts
The "target bucket" and "target bucket reduced" names are a bit opaque and unclear. The "archive bucket" is the location where archiving is saved to and loaded from. The "public bucket" is the bucket where data will become publically available. These names are clearer and help to explain where and why certain files will be archived there.