Add control over whether an alerts baseline is saved as the reference
Often when we run the new baseline workflow, it will be to create a new alerts baseline reference for daily alerts. In this case we want to archive the result in a central location so it can be used by daily workflows.

In other cases, we might want to run the alerts baseline just to get some diagnostic data. In these cases, we do **not** want to overwrite the existing reference in the archive.

This is now controllable using the ALERTS_BASELINE_SAVE_REFERENCE env var in the archive script.