Fix do_ICs codepath incorrectly referencing ncoutb
ncoutb will only exist if do_BCs is true, which is not necessarily the case in this block. This seems likely to be a copy/paste error, although it's odd that it wasn't caught until now. If do_BCs and do_ICs are both true, would this cause a bug with the initial conditions?