Update posterior emissions format based on new prior output
This change includes corrections to the existing implementation which were discovered as part of determining the new format for the prior. The "lat_bounds" and "lon_bounds" variables have been removed, as they were not CF compliant (CF only supports bounds on coordinate variables). The extract_bounds implementation also had a bug which caused emissions to be shifted southward a significant degree when visualising results.