Update alerts format based on new prior output
Similar to the previous commit, this change includes updates to the alerts baseline and alerts output format based on recent changes to the prior output. Although alerts do not directly rely on output from the prior, the prior is currently used as a proxy for the domain file, since it includes many CF conventions and quality of life updates.

These changes to alerts require that a prior output file is used as the "domain file" input, but this is already the case in production. In the future when the domain format is updated this will no longer be necessary.