Re-introduce vertical dimension in posterior output
The coordinates for the ch4 and prior_ch4 variables in the posterior output had previously been reduced to just time, y, x since we don't currently do any calculation of emissions at different vertical levels. However, this may be something we want to do later, so this leaves a dimension in the format to allow us to add vertical layers later.